chore: remove unused components and rename logger module for better organization
