+++ src/extension/extensionLogger.ts
